Introduce customizable error descriptions.
The original one is the default, which is InlineErrorDescription.
In the command line, use -i, or --inline.

The second one would try to put the expected result and actual
result with newlines, which is much easier to tell the difference
with human eyes when the inspected string is quite long.
For example, try to test a markdown formatter with a long text.
In the command line, use -l, or --newline.

The last one tries to put this further with external diff command.
Which might not be as portable as the one in minitest, but could be
already useful on most of the platforms other than Windows.
In the command line, use -f, or --diff.

Just try to be consistent with formatter argument, we also provide
--error which you could specify Inline, Newline, or Diff, or even
user provided error descriptors. However I failed to find a letter
for this argument. I guess providing --error might be good enough.