Make helper methods with blocks to work.
Sometimes it is convenient to use a shared helper method which yields to a block (for example, in my case, it was a  method which takes a string and yields the string several times, each time with single character changed to something else). Unfortunately, Bacon currently doesn't support such helper methods defined within Bacon::Context instance and they have to be defined directly in Bacon::Context or in Object, simply because the blocks are not passed at all to the parent context.. The change above tries to fix this.